+++ framework/scenario/src/debug_executor/static_var_data.rs
